Bug 1445106 - don't try to have a feed reader in a frame, r=mak
MozReview-Commit-ID: 24EESnD4UaT

+++ browser/components/feeds/FeedWriter.js
@@ -64,6 +64,9 @@ function convertByteUnits(aBytes) {
   return [aBytes, units[unitIndex]];
 }
 
+XPCOMUtils.defineLazyPreferenceGetter(this, "gCanFrameFeeds",
+  "browser.feeds.unsafelyFrameFeeds", false);
+
 function FeedWriter() {
   this._selectedApp = undefined;
   this._selectedAppMenuItem = null;
@@ -778,6 +781,9 @@ FeedWriter.prototype = {
   // BrowserFeedWriter WebIDL methods
   init(aWindow) {
     let window = aWindow;
+    if (window != window.top && !gCanFrameFeeds) {
+      return;
+    }
     this._feedURI = this._getOriginalURI(window);
     if (!this._feedURI)
       return;
@@ -877,6 +883,9 @@ FeedWriter.prototype = {
   },
 
   close() {
+    if (!this._window) {
+      return;
+    }
     this._document.getElementById("subscribeButton")
         .removeEventListener("click", this);
     this._handlersList
@@ -895,7 +904,7 @@ FeedWriter.prototype = {
   },
 
   _removeFeedFromCache() {
-    if (this._feedURI) {
+    if (this._window && this._feedURI) {
       let feedService = Cc["@mozilla.org/browser/feeds/result-service;1"].
                         getService(Ci.nsIFeedResultService);
       feedService.removeFeedResult(this._feedURI);
@@ -904,6 +913,9 @@ FeedWriter.prototype = {
   },
 
   subscribe() {
+    if (!this._window) {
+      return;
+    }
     let feedType = this._getFeedType();
 
     // Subscribe to the feed using the selected handler and save prefs
